Separating service factory + auth intercepter from the apiService
