+++ app/src/main/java/pl/llp/aircasting/data/api/services/AuthenticationInterceptor.kt
@@ -1,0 +1,15 @@
+package pl.llp.aircasting.data.api.services
+
+import okhttp3.Interceptor
+
+class AuthenticationInterceptor(private val authToken: String) : Interceptor {
+    override fun intercept(chain: Interceptor.Chain): okhttp3.Response {
+        val original = chain.request()
+
+        val builder = original.newBuilder()
+            .header("Authorization", authToken)
+
+        val request = builder.build()
+        return chain.proceed(request)
+    }
+}
